passing reset_params test
