+++ dfdx-macros/Cargo.toml
@@ -27,6 +27,8 @@ proc-macro = true
 proc-macro2 = "^1.0"
 quote = "1"
 syn = { version = "^1.0", features = ["full"] }
+darling = "0.14"
 
 [dev-dependencies]
 dfdx = { version = "0.10.0", path = "../." }
+rand = { version = "0.8.5", default-features = false, features = ["std_rng"] }
